Add token revocation for existing email confirmation tokens and fix capacity type in Property validation

+++ backend/Guestly.Domain/Entities/Properties/Property.cs
@@ -177,7 +177,7 @@ public class Property : BaseEntity
     private static void ValidateProperty(
         decimal pricePerNight,
         decimal cleaningFee,
-        decimal capacity
+        int capacity
     )
     {
         if (pricePerNight <= 0)
